+++ src/main/java/com/example/cat/DAO/CoursesRepo.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.cat.entity.Courses;
 
-public interface CoursesRepo extends JpaRepository<Courses, Integer> {
+public interface CoursesRepo extends JpaRepository<Courses, Long> {
 
 //	public Courses findByName(String name);
 }
