+++ src/main/java/com/example/cat/DAO/StudentRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.cat.entity.Student;
 
-public interface StudentRepo extends JpaRepository<Student, Integer> {
+public interface StudentRepo extends JpaRepository<Student, Long> {
 
 }
